+++ solum/include/solum/vm.h
@@ -65,6 +65,7 @@ struct SolVM {
     SolObject *block_class;
     SolObject *array_class;
     SolObject *string_class;
+    SolObject *object_class;
 
     bool had_error;
 };